final project/website/requirements.txt
bs4 ==4.11.1
flask==3.0.0
fuzzywuzzy==0.18.0
numpy==1.23.5
pandas==1.5.3
scikit-surprise==1.1.3
selenium==4.15.2
tqdm==4.64.1